+++ resources/release-notes.md
@@ -8,8 +8,10 @@
 - (Bartington, HelioSwarm, IMAP) Increase app area size by 30 %
 - (Bartington, HelioSwarm, IMAP) Exported results are named after the selected mission (e.g., `imapAnalysis` instead of `analysis`)
 - (IMAP) Show any available metadata for sensor setup
+- Add UI tests for Bartington, HelioSwarm, and IMAP
 
 ## Software
 
+- (IMAP) I-ALiRT with only one sensor is still valid
 - (IMAP) Fix issue when no acknowledge or completion events are available in logs
 - (IMAP) Fix issue in detecting transition events when no science events are detected